fix: Move InsuranceFundAccount to state
